backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Pillow==10.3.0
ImageHash==4.3.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
